dev_requirements.txt
# Development requirements for wagtailsvg
# Install with: pip install -r dev_requirements.txt

# Core requirements
wagtail>=6.0,<7.0
django>=4.2,<6.0

# Development tools
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0

# Optional: For even faster development with auto-reload
watchdog>=3.0.0 